topcat: fix session synthetic column bug
A bug was reported by Bertrand Goldman: if you had a synthetic column
that had been loaded from a session file, and renamed it in the
columns window, the name in the columns window didn't get updated
(though the name in some other places did).  This came down to the
fact that the SyntheticColumn object was working with a copy of
the ColumnInfo metadata object rather than the original, because of
how synthetic columns are reinstated into the TopcatModel by the
session decoder (TopcatCodec2).  I made a change that fixes the
reported bug, and I think improves matters generally, but this
confusion between use of copies or mutable objects is nasty,
and it's possible there are some other related bugs there somewhere.