topcat: use PolygonPanel in StackPlotWindow
The plot window now offers an action that lets you define a
subset positionally by clicking at the vertices of a polygon.
The subsets thus defined are defined in terms of algebraic
expressions based on the positions identified by the user,
so suitable for export.

This only works for PlanarSurfaces, i.e. mainly the PlanePlot
(also the TimePlot).