vo: TapCapabilitiesDoc extract non-TAPRegExt metadata
The TapCapabilitiesDoc class now records information about the
standard endpoints and capabilities listed in the TAP
capabilities file, as well as the TAPRegExt-specific items,
using a new type, StdCapabilityInterface.

At least some of this new StdCapabilityInterface information
is required in TAP 1.1 to locate all the service endpoints
along with their associated securityMethods.

This implementation is based on the current draft of TAP 1.1,
still under discussion.