ttools: avoid using URL as Map key in UrlMocCoverage
I just found out that URL.equals does hostname resolution,
which makes using a URL as a Map key a very bad idea,
since the resolution operation is in principle network-dependent,
so could be quite slow.

Adjust the code to use a String as the Map key instead.
Benchmarking doesn't show a significant change in execution time,
but it's safer this way.