ttools: avoid using Set<URL> in taplint ExampleStage
I just found out that URL.equals does hostname resolution,
which makes using a URLs as the contents of Sets a bad idea,
since the Set addition operation requires equality determination,
and the resolution operation is in principle network-dependent,
so could be slow.

Adjust the code to use String as the set content type instead.