topcat: BasicAuth popup not parented by ControlWindow
The popup window used for HTTP Basic Authentication previously had
a parent component which was the main topcat ControlWindow.
That sounds sensible, but isn't really; it meant that the ControlWindow
was getting pulled to the top of the window stack when Basic
Authentication was required, which is not really helpful.
So the parent window is now null.  That probably just means the
popup will appear in the center of the screen, which is reasonable.