ttools: modify stilts invocation logging sequence
Change the sequence in which some of the configuration tasks
are done during stilts command setup.  Previously, some of the
configuration stages which emitted logging messages were
performed before the configuration of logging itself.
That meant that logging messages could get output that shouldn't be,
or vice versa.  Add a callback to the LineInvoker.invoke method
for the relevant code to fix this.