vo: rename EndpointSet->TapService
The class that aggregates the service endpoints for a TAP service
is renamed from EndpointSet to TapService.  This is a better name
partly because it includes the term "TAP".
However, it's also going to be necessary to add some content to
it alongside the endpoints themselves, for instance information
about TAP protocol version.  Given that, EndpointSet would have
been an unhelpfully inaccurate name to give it.

The behaviour at this refactoring commit however is not changed
at all.