ttools: taplint check TAP 1.1 size/arraysize consistency
The "size" and arraysize columns in the TAP_SCHEMA.columns table
are tested for consistency according to TAP 1.1, where applicable.

This implementation is a bit scrappy.  The column testing is done
in two separate places for no good reason.  This really ought to
get done in the LintTapSchemaInterrogator, and/or the testing done
there should be moved elsewhere.  But it works.

These changes got reworked a few times following clarifications
and feedback from the TAP 1.1 author and implementors.