ttools: modify taplint capabilities doc read handling
The TAP capabilities document is required for several of the
taplint stages.  Previously, it was read by the CAP stage
(CapabilitiesStage) and acquired from there by other stages
that need it.  But for a couple of reasons, it's better to
centralise the capabilities reading outside of any of the
stages, so do that instead.  Also extend the CapabilitiesHolder
interface to provide standard interfaces as well as the
TAPRegExt information available from the capabilites document.