util: fix Base64InputStream bug
There was a bug in Base64InputStream.  If a Base64 stream ended with
a single padding character ("=") to make it up to 4N output bytes,
then doing a second read after the first one that encountered the
stream end would result in an EOFException (Unexpected end-of-file).
This is because the state machine's state did not get updated once
the padding sequence had begun.  Although this class has been in use
here for a long time, this bug had gone unnoticed evidently because
the stream was never re-read after the end of file was encountered.

Fix it.