topcat: invocation scripts look in OSX /Applications dir
On MacOS, the topcat and stilts invocation scripts now add the absolute
path /Applications/TOPCAT.app/Contents/Resources/Java/topcat-full.jar
to the list of places they look for jar files.  Apparently this
is required for Homebrew installations, since they put the scripts
in /usr/local/bin, but the application in /Applications.

This fix applied on the advice of Paul Price (Princeton).