ttools: remove duplicate INFO/@name votlint warning
The VOTable validator (votlint, also used in taplint) warns if
there are multiple elements of the same type with the same @name
attribute in the same scope (same parent).
But Markus Demleitner argues that this is inappropriate for
INFO at least, since it's reasonable to have several INFOs
with the same name (not least QUERY_STATUS).  So remove that
check for INFOs.