topcat: adjust startup script location in OSX bundle
For the OSX bundle that goes to make the DMG file,
duplicate the topcat and stilts startup scripts from the /bin
directory so they appear in /TOPCAT.app/Contents/Resources/bin
as well.

This follows a request from the Homebrew people, who want the
scripts under /TOPCAT.app so they can be symlinked from somewhere
or other.  I took advice on exactly where the scripts should go,
and got several different answers (/TOPCAT.app/bin sounded logical
to me, but apparently not).  This one is at least OK for
Homebrew, and seems reasonably unobjectionable.

They probably don't need to be in the top level /bin directory now,
but since they've lived there a long time, and people may be relying
on them, leave those in place for now.