votable: fix exception configuration bug
I was setting the Throwable cause of a SAXParseException twice
(to the same value) once in the constructor, and then with a
separate call to initCause.  This had gone unnoticed, but apparently
(report by Ole Streicher from Debian build) it causes a runtime
failure in SerializerTest at JDK11.  Now fixed.