topcat: add Algebraic Subset From Visible action
Add a new action that can define a new subset from the visible
region, but using a JEL expression rather than just flagging
the currently visible points using a bitmask.

This is a bit more work, but in general more useful, since the
expression can be exported elsewhere, like the polygon option.
In 2d, polygon could be used just as well, but in the cube plot
it could be quite useful (e.g. the Hyades example).