topcat: add query status information in Control Window
The "Rows" line in the control window properties panel now displays
information extracted from any non-OK INFO/@QUERY_STATUS nodes
in the input VOTable (if the input file is a VOTable) - see DALI.

This gives the user an indication that, e.g. there was an overflow
or (non-fatal) load error when the table was loaded.  Otherwise they
would have to go poking through the Parameters list, with some
knowledge of DALI.

I should have done this years ago.