topcat: grey out not erase bad syntax in Function window
When you enter an expression in the Function Expression field of
the Function layer control, it used to delete the text to show you
it was wrong.  Now instead it just greys it out, which means you
can edit it rather than having to start from scratch each time,
which is a significant usability improvement if you're using this
window.

Implementation is not very pretty (which is why I haven't done it
before now), because the validity of the field depends on
interactions between different configuration items, and the
user has to be informed of the error at some suitable point.