cdf: fix bug in global attribute handling
When constructing a StarTable parametem from a CDF file
Global Attribute, an attempt is made to turn the (potentially
inhomogeneous) set of AttributeEntries into a sensible scalar
or array object.  There may be null items in the list of entries,
but the code was not taking account of this.

That caused a NullPointerException, which didn't halt processing
(it got caught), but it did prevent the GlobalAttribute from
getting turned into a table parameter, with a
"WARNING: Omitting complicated global attribute" log message.
Fix the processing so that null AttributeEntry values are just
ignored.

That's fine, but it was treating null entries as