topcat: prevent recursive JEL expressions
Take steps in the places where synthetic columns or subsets can
be redefined that they do not allow expressions that directly
or indirectly reference their own name.

If such self-references occur, the expression can't be evaluated
and some kind of infinite loop or stack overflow results.
This has always been possible, but not very common; it may become
more common for subsets with the newish graphical synthetic subset
definition actions.

This code may not be bulletproof, but it should prevent the large
majority of possibilities of this happening.