ttools: avoid underflow in HistogramPlotter bins
HistgoramPlotter now uses a point in the middle of the axis rather
than at the lower end of the axis as the "representative point"
to pass to the BinBag constructor.  This will normally make no
difference, but near the end of the IEEE754 representation range
(Double.MIN_VALUE) it can prevent underflow problems.