ttools: avoid over/underflow in some calculations
Rewrite the algebra in a number of places, replacing log(a/b)
with log(a)-log(b), to avoid numeric overflow (Infinities)
in the case that a/b is very large, most likely because either
a or b is near the limit of IEEE754 representation
(i.e. near Double.MIN_VALUE or Double.MAX_VALUE).

This was stimulated by a user trying to plot histograms of
temperatures of about 2e-323(!), which was causing infinite
loops, assertion failures, stacktraces, invisible plots etc.
Those histograms and KDEs now work in tests, but there may be
other instances of the same kind of thing that ought to get
fixed up elsewhere in the code.

These numerical rearrangements give rise to some tiny changes
(mostly just a couple of pixels) in a few of the example plots.