table: redefine StarTable.getColumnAuxDataInfos with generics
Redeclare the return value of StarTable.getColumnAuxDataInfos
as a List<ValueInfo> rather than an unparameterised List.

This intention has always been in the documentation
(the class was specified before generic types existed in the
Java language), so it's not expected to cause problems,
but it is a change to the API of the central StarTable class,
so some issues could occur.