fits: add HealpixFitsTableWriter
Provide a special-purpose table output handler specifically for
writing HEALPix-format FITS files.  This shares most of the same
behaviour and implementation with the standard FITS output handler
(which also inserts the relevant HEALPix headers), but it's more
picky: if presented with a table that's not a HEALPix map, it
will throw a TableFormatException, and it also rearranges columns
for maximum HEALPix-FITS compatibiilty if required.