ttools: fix Sky readRanges to honour extended coord ranges
The SkySurfaceFactory implemention of readRanges looks at the
coordinate ranges but did not, unlike the other gometries,
call the extendCoordinateRanges methods of the submitted layers,
so it failed to auto-range correctly layer types that provide
a non-blank implementation of that method.

Fix it so that it does use that method, by calling the PlotUtil
utility method used by the other SurfaceFactories.
Implement it in such a way that it doesn't keep ranging after it
knows the full sky is covered.