ttools: improve HealpixPlotter coordinate handling
Rework HealpixPlotter so that its healpix index coordinate is
declared as a 'positional', rather than an 'extra' coordinate.
This means that the rest of the plotting system understands
how data points map to the plotting surface (in this case, the sky),
so that for instance auto-ranging and interactive point selection
work without any extra effort; previously, they just didn't.
A new DataGeom implementation, HealpixDataGeom, is required to
define this coordinate mapping.

As far as the plot2 system itself goes, that works fine.
However for interacting with the UIs, there are some things
elsewhere that need to be configured specially to make this work.
The problem is that the TOPCAT and STILTS UIs assume that
positional coordinates are of (more or less) the same type for
all plot layers in the same plot type; in this case, that
all positional coordinates used in the Sky Plot are (lon,lat) pairs,
which is not the case for Healpix.  The form this assumption takes
is that (more or less) a single DataGeom is used for each plot type,
in this case a SkyDataGeom.  But HealpixPlotter needs a HealpixDataGeom.

The right thing to do would be to rework the UI implementations
so that the DataGeom type is specific to each plotter.  However,
that's probably a complicated and disruptive job, and at the moment
HealpixPlotter is about the only instance where it's required
(though maybe the Histogram-like plots are trying to do this too?)
This means that (a) it's a lot of work for a single exception to a rule
and (b) it's not really clear how best to do this generalisation.

So, for now handle the HealpixPlotter in topcat and stilts as a
special case, and provide custom handling to supply the right
DataGeom when it is encountered.  This seems to work fine.
If other instances of the same problem crop up in the future
(different positional parameter configurations in a given PlotType),
revisit this solution and maybe come up with something more
general and respectable.  If that happens: I think maybe the method
PlotType.getPointDataGeoms, which looks a bit questionable anyway,
is completely misconceived and should be replaced by something else.

Falling out almost without effort from this is raising the
HealpixPlotter maximum HEALPix level from 13 to 20, the
maximum supported by the PixTools HEALPix library.
The only thing that really needed changing was making
HEALPIX_COORD (now defined in HealpixDataGeom not HealpixPlotter)
into a LongCoord instead of an IntegerCoord.
A comment in the existing code suggests that it ought to be
more complicated than that, but it seems to work OK,
e.g. for moderate sized (not all-sky) files at level 16.
Probably some changes elsewhere in the code have made this easier
since I wrote that, though I can't immediately see what.
Maybe it's because I was just thinking about all-sky
high-level maps before, which would be challenging still.
Hopefully I'm not missing something important.