topcat: initialise plot with Healpix where appropriate
In the case that the current table can be identified as a HEALPix
index table, and a sky plot is requested, initialise the plot
with a HEALPix layer rather than the usual scatter plot layer.

This special casing is a bit questionable (might confuse users),
but it's pretty likely that this kind of table will not have sensible
lon/lat coordinates, so a normal scatter plot is not wanted.