ttools: add generics in various places
Mostly change the argument type of Filter.createStep from from
raw Iterator to Iterator<String>, but tidied up various other
generics in the process.  The actual values dispensed from
Iterators used in this way should always have been compatible
with this type, so this is just a case of formalising in the
source code the contracts that were always in place.