util: avoid use of sun.* class in BasicFileChooser
The class uk.ac.starlink.util.gui.BasicFileChooser relied on the
undocumented class sun.awt.shell.ShellFolder.  This questionable
behaviour caused build problems on certain platforms (e.g. Java 9),
and was flagged as a starjava github Issue since 08/2017.

This commit removes references to the undocumented class, so that
the build problems (and build warnings) disappear.  It also means
that the class no longer does what it was supposed to, viz. work
around a bug in making sense of Windows .lnk files.  That workaround
may(?) be no longer relevant given the changes in the java and
Windows platform in the 15 years since it was adopted.

Code which uses this class (SPLAT) should therefore probably just
revert to use of a standard JFileChooser instead.  Alternatively
it would be possible to reinstate the old behaviour using reflection.

If all client code (SPLAT) relinquishes use of BasicFileChooser,
that class can be removed from starjava.