ttools: add custom doc auto-generation for Healpix layer
Because HealpixLayer does not use the SkyPlot's standard DataGeom,
the autodocumentation needs special casing for Healpix.
Add some ad-hoc code that makes it work.

This is messy, but it's not clear how best to generalise it.
If other instances of the same thing crop up elsewhere,
reconsider the implementation.  See also f01f6e601b3841.