topcat: BasicCoordLayerControl respects Current Subset
The GUI in the BasicCoordLayerControl (this currently affects Healpix
and Spectrogram layer controls) now respects the currently selected
Row Subset; the subset selector now defaults to the TopcatModel's
current subset rather than RowSubset.ALL.

This is what it should do, and especially important for Healpix plots
with very large tables, since they are currently displayed in a way
that takes a lot of memory if the whole of a large dataset is plotted.

This fix was pointed out by Heinz-Bernd Eggenstein.