ttools: add and use AdaptiveBinList implementation
Add a new implementation of BinList that starts off as hash-based
but switches to array-based if more than a certain proportion
of the hash bins get filled up.  This is more flexible and will
cope with larger bin lists.  In practice, it allows larger
Healpix maps to be plotted (HealpixPlotter) without incurring
(much) extra cost for non-large ones.

A few changes to the other BinList API and implementations went
along with this.