topcat: allow user edit of Shape metadata in Columns Window
The ColumnInfo shape field in the ColumnInfoWindow can now be
edited by double-clicking on it, in the same way as other metadata
items like UCD and Name.  I'm not sure why this wasn't enabled
in the first place.

It is mostly useful to set a fixed length for array columns that
have unknown length.  This may be required when preparing tables
for output, and it's also very helpful if you want to Explode an
array column, since that action is not enabled if the column length
is unknown.