ttools: SkySurface can now draw scale bar
The SkySurface now has the option to draw a little bar annotated
in degrees, minutes or seconds at the bottom left of the plot,
to make it easy to see the plot scale.  At this commit, the UI
does not provide controls for this option, so it is switched off.