ttools: add controls for plotting sky scale bar
The STILTS plot2sky command and the TOPCAT SkyPlotWindow now have
the option to draw a scale bar at the bottom of the sky plot.
The plot2sky parameter is called scalebar.

This is currently switched on by default; I'm not sure if that's
a good idea, but otherwise most users won't notice or use it.

One could add several more configuration options
(approximate size, foreground and background colour etc),
but I don't want to inflate the UI with more additional
complication than is necessary.