ttools: add SurfaceFactory.getPlotMetric method
SurfaceFactory implementations can now supply a PointMetric object,
which knows how to describe the distance(s) between two points on
the graphics surface.  This can be used to provide distance
measurement widgets in the GUI.

Non-null implementations are provided for Sky, Plane and Time surfaces.
It can't be done for 3d plots.