ttools: add Color parameter to LineTracer.addVertex
The LineTracer now (like LineTracer3D) has a Color argument to its
addVertex method, so it can paint lines that vary in colour over
its length, rather than just being of a single colour.
This generalisation required some rewriting, and results in a few
pixel-level changes in the example plots.  Although the algorithm
is now somewhat more complicated, for lines that are actually
of a single colour, the performance is not expected to degrade much.

At this commit, none of the plotters take advantage of this
capability.