ttools: add Projection.isContinuousLine method
New method in (sky) Projection class that can determine whether
a line between two screen points includes a discontinuity in
the sky projection.

Current implementations are correct, but the SkyviewProjection
implementation only works if the discontinuity is down the
longitude=+/-PI line.  That is the case for all discontinuous
SkyviewProjection instances currently in use, but one could
imagine instances that have discontinuities elsewhere, so
make some adjustments in the documentation and visibility of
SkyviewProjection creation methods to avoid inadvertently
falling foul of this restriction.