ttools: fix ShapePlotter to work with null DataGeom
Some generalisations so that ShapePlotter works with a null DataGeom,
which is what you get if you have no positional coordinates.
It's questionable whether using it like this is a good idea.

For instance, FloatingCoordAuxReader will now still function, but it
can't restrict the values it reads to those that are currently visible.