ttools: GlyphPaper users can now use 3D painting
PixOutliner and ShapeMode, that both paint to GlyphPaper to get
the bitmap of a glyph, now use the 3D painting method if the
paper is 3d rather than just using the 2D method in any case.
The effect is likely to be similar (since the Z coord is ignored)
but it may not be identical, for instance 3d plotting may
clip output according to the Z coordinate.