topcat: improve coordinate defaulting in plot windows
Make some implementation changes so that coordinates in plot
windows default to hopefully more helpful values when the
plot layer is created or the table is changed.

In particular, for multi-position plots (e.g. sky position pair plots),
multiple sets of coordinates will be filled in if multiple sets
of likely-looking coordinates can be found in the table - e.g.
two sets of coords with pos.eq.ra/pos.eq.dec UCDs.

Also, the Plane and Cube plots now look for XY/XYZ coordinates
and use them as defaults if they can be found (by UCD or column name).