ttools: add CAR0 sky projection
Add a new sky projection option, Car0, which gives you a Plate Carree
projection the same as Car, but centered on longitude 180 rather than
on longitude 0.  This is requested by Batiste Rousseau, since it's
more natural for the planetary community (possibly others too).

This required some reworking of the SkyviewProjection class, which
was a little bit of a mess in any case.  It's better now.