ttools: rework AuxScale ranging
Adjust the way that Auxiliary (non-positional) range determination
is done for more flexibility.  Instead of a mutable plot.Range object
that is updated by plot layers with information from data, there is
now a Ranger interface which does the ranging and produces an
immutable Span object.  The point of this is so that different
Ranger implementations can do different kinds of ranging.
The target is histogram-like ranges that can do non-linear value
scaling based on the details of data value distribution,
rather than just looking at the extreme values.

At this commit, positional ranging is not affected.  It could be
(and maybe will be at some point in the future) updated in the
same way to avoid use of the old plot.Range class and for consistency,
but it's not so pressing, since data-based positional axis scaling
is probably not useful.