ttools: Scaler is more self-documenting
Add some methods to the Scaler interface so that it contains all
relevant information about the scale it describes.
This object can then be used, rather than the Scaling+Span combination,
when constructing ShadeAxis instances, which makes the interface
and implementation of ShadeAxis a bit clearer.

Scaler is also now marked @Equality.