fits: defend more carefully against long header strings
When constructing output FITS headers, some card values come from
strings of uncontrolled length, e.g. TCOMM.  These are written
if they are short enough, and just discarded if they are too long.
Until now, this was done mainly by trying to add them to the header,
and relying on nom.tam.fits to throw a HeaderCardException if it
failed because they exceeded the maximum length enforced by FITS
header card structure (68 characters).

However, in some more recent versions of nom.tam.fits the extended
header convention can optionally be supported
(nom.tam.fits.FitsFactory.isLongStringsEnabled(), not set by default).
That means that no HeaderCardException is thrown, and the overlong
keyword is added.  This can cause trouble when reading the headers
back in (I haven't identified exactly where that happens; it shouldn't
be a problem since extended headers ought to be invisible under
the backward compatibility arrangement, but I have reports from
STIL users at IPAC which look like this is happening somehow).

This commit tests explicitly for overlong string values, and does not
attempt to add them to the header.  It ought to fix the problems.