ttools: define impl-specific Scaling sub-interfaces
The way that a Scaler is created from a Scaling instance is now
defined by implementing one of the Scaling sub-interfaces.
Currently there's only one of these, Scaling.RangeScaling,
but when histogram-based scaling is introduced a new one will
be necessary.

In view of the additional complexity this arrangement generates
in the code, Scaling has been redefined as an interface instead
of a class, with the implementation code in a new class Scalings.