vo: extract TapService instances from TapCapabilitiesDoc
Provide methods in class TapServices to pull out a list of TapService
instances from the content of a parsed TAP capabilities document.

There are various ways of doing this: the current implementation
pulls out one TapService for each interface with
standardId of "ivo://ivoa.net/std/TAP", which corresponds to what
will probably be in the TAP 1.1 standard, though that will be
modified by declared securityMethods.