topcat: minor improvements to CDS Upload XMatch window
Added a bumper to the Vizier table selection combo box, and adjusted
the layout so the selector works better for wide and narrow windows.

Also fixed the MOC download metadata population code so that it
doesn't make the same request twice every time in quick succession.
This may be more necessary following recent changes to the Downloader
implementation (seemed to be unreliable before this edit).