vo: add missing generics to LinkedBlockingStack
Add some parameterised types where they were missing to the
LinkedBlockingStack implementation, so that there are no
compile-time unchecked warnings.

I'm not 100% certain that I've done this correctly where the original
authors didn't, but I think it's OK, and the methods affected are
not ones that I'm using anyway.  As a precaution, I've demoted the
visibility of this class to package-private; it's just used as
implementation code for TapServiceKit, it's not intended as part
of the public API.