topcat: work round ListSelectionModel performance issue
Heinz-Bernd Eggenstein reported the following issue: on OpenJDK 1.8
in Debian Stretch, if you open a large (500krow) table, open the
table viewer window, and create a large subset ($0<250000),
the GUI hangs for many seconds.  This doesn't happen in e.g. my
Oracle java development environment.

This turns out to be down to the fact that a ListSelectionEvent is
getting sent and acted on for each selected row, even though
isAdjusting is set true on the corresponding ListSelectionModel.
I don't think that should happen, though I'm not certian of the
wrongs and rights of it ... anyway I have overridden the relevant
event firing method in DefaultListSelectionModel so it doesn't
happen any more.  I don't expect any negative consequences, but
I'm changing the behaviour a bit from the JRE behaviour, so it's
possible something bad will happen.