ttools: use (IVOA-validate) User-Agent header
Tasks taplint and datalinklint now use the vo.UserAgentUtil class
to insert validator marker tokens into all the HTTP requests they make.
This suggestion was discussed in the IVOA Operations Interest Group
in late 2018, for the convenience of services that want to identify
validation accesses.

This is a little bit unrespectable, since JVM-wide behaviour is
being influenced by running STILTS tasks, and in principle might
disturb activities of other processes in the same JVM.
In practice, it's not that likely that these will be run in JVMs
that are doing other things at the same time, but in any case
some attempt is made to clear up and restore the original state
at task end.