votable: expose referenced TIMESYS cols as aux metadata
For FIELD elements that successfully reference a TIMESYS element,
the relevant information is now made available from the
column aux metadata.  This means the TIMESYS information is now
accessible via the StarTable interface, and can therefore be
seen from TOPCAT's columns window or STILTS's column metadata
interrogation facilities.

This works just the same as for COOSYS.  Like COOSYS, it only
works for FIELDs (columns) not PARAMs, because while ColumnInfo
has hooks for aux metadata, ValueInfo does not.  This is probably
a flaw in STIL design.