ttools: avoid Locale-specific axis annotation in SkySurface
In the Sky Plot when grid labelling was set non-sexagesimal,
the grid labels were being generated in a way that was sensitive
to Locale; if you ran with -Duser.language=fr, the decimal
separator was a comma (",") not a period (".").  In a related bug,
line labels with no fractional part appeared with a trailing comma
(e.g. lines were labelled "30," rather than "30" in ","-using locales).

Prevent both things (my guess is that even e.g. French topcat users
would rather see "." as a decimal separator).  This required
a small hack to the source code of GridLiner, which is copied from
SkyView to do the sky axis labelling.  Since I don't want to mess
with SkyView innards, the hack may not be bulletproof, but
I think it will do the job.