topcat: report JEL/ADQL expressions for graphic subsets
When the user has completed drawing a Figure or graphically specified
an axis range for definition of a subset, the subset addition dialogue
now displays some additional information about the figure so defined:
the generic JEL and ADQL definitions of the region.

This can be very useful information, especially the ADQL which the
user can pull out to paste in to a TAP query representing the same
region.  However, this functionality is not very well advertised,
and not in a place where the user is likely to go looking for it,
so it's of limited use in practice.  This bit of GUI is therefore
considered provisional; the idea is to support geometric regions
in some more prominent fashion in the future, at which point this
functionality will probably be moved there.