ttools: update taplint for ObsCore-1.1 Erratum #1
Changes to UCDs for ObsCore-1.1 Erratum #1.

The UCDs for obs_publisher_did, publisher_id and o_stat_error have
been updated.  If the pre-Erratum UCDs are detected in the ObsCore table,
they are now flagged as an error, but the message explains that they
have been updated by an Erratum.

UCDs for dataproduct_type and dataproduct_subtype,
present in some versions of the Erratum process, have not been
changed here since they shouldn't be in the Erratum
(this was agreed at TCG F2F 12 May 2019).

Believed working but not yet tested on TAP services.