vo: upgrade adqlLib to VOLLT 1.5
Gregory Mantelet's ADQL parsing library, with the new branding "VOLLT"
at this version, is upgraded from 1.4 to 1.5.  This should result in
a few minor improvements to ADQL parsing, hence syntax highlighting
in the TAP window.  I think some of the error messages are improved
as well.

This version also provides a new function ADQLParser.tryQuickFix
which might be a nice option to offer the user, but at this commit
it is not used.

The library is currently hosted at https://github.com/gmantele/vollt.
The jar file in this commit was rebuilt by me from source with
very minor changes to get a Java6-compatible version.