vo: adjust manifest of adqlParser.jar library
The MANIFEST.MF for adqlParser.jar included "Class-Path: .".
This is bad for standalone-ness in general, and in particular
interfered with the topcat build-standalone build target.
This version of adqlParser.jar should be exactly the same as
the previous one (that is, adqlLib 1.5 with very minor build
tweaks) apart from the manifest file.