topcat: add pause button to Activation Window
The activation window allows you to schedule a sequence of activation
actions, one for every row in the (current subset of) the selected table.
There was already a button that can abort this sequence at user request,
but this commit adds another button that can be used to pause the
sequence, and restart it from the same place later.

This was added at the suggestion of Petr Skoda.