topcat: activation sequences update row highlights
The activation sequences that can be set off from the Activation Window
(activate all rows...) now make sure that the row they are currently
activating is updated elsewhere in the applcation (row highlighted in
table viewer and plots) as they go along.

Thanks to Petr Skoda for pointing out this is what should happen.