ttools: fix regression bugs in pixsample command
The HealpixTableInfo processing introduced in the last release
changed the way that FITS headers relating to HEALPix map encoding
made it into StarTable objects, resulting in the scheme ordering
information being lost.  That caused trouble in the pixsample command,
since it looked (more or less) directly at the FITS header metadata.
Rework that code so that it looks at the HEALPix header information
corresponding to the uk.ac.starlink.table.HealpixTableInfo class.

Also change the default of the pfmt parameter from FITS to (auto),
since with the HealpixTableInfo changes it's now feasible to store
this information in VOTables etc.

Note the pixsample command could be working harder here, to work
with partial instead of all-sky maps.  Leave that change for a
future date.