ttools: add Use Position toggle to Polygon plot form
You can now choose if you want to use the row reference position
as one of the entries in the polygon vertex list or not.
Being able to exclude it makes it easier to plot polygons
where the vertices are already available as an array but
the reference position is not one of the vertices.
Now you can just toggle UsePosition off, rather than having to
supply the first entry as a row position and entries 2..N for
the vertices.  Should have done this when the polygon form was
first introduced.