topcat: note alternative TAP_UPLOAD notation in SUN/253
Add a note in the discussion of table uploads in the TAP window
that you can use the table label as an alternative to T<ID>
in the TAP_UPLOAD specification for loaded tables (subject to
label syntax constraints).

I didn't know (had forgotten) that this undocumented feature existed,
but Stephane Erard was using it in an example he was having (other)
problems with; since it works, it might as well be documented.