fits: permit zero-length string columns
When writing a FITS character column in which all the entries
are blank (zero-length strings), write it as a zero-length
character array.  Previously it was written as a single-character
column, annotated with the comment "zero length strings are
probably a bad idea for FITS output".  However, the FITS standard
explicitly says that repeat counts can be zero, so withdraw this.

At the same time delete warnings/errors in VOTable read processing
corresponding to zero elements of the arraysize specification.

The fact that round-tripping all-null VOTable columns with
datatype="char" arraysize="*" via FITS ended up with arraysize="1"
was causing problems in some TAP upload queries.
This should fix that, the unit tests still pass, and I can't see
any reason why it's going to cause problems.  However, this is a
change to code that's been in place for a long time, so look out
for issues.